requirements.txt
click>=8.0.0
pyyaml>=6.0
gitpython>=3.1.0
google-genai>=0.8.0
anthropic>=0.40.0
rich>=13.0.0
pathspec>=0.11.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0